api/sklearn_worker/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
joblib
boto3
botocore
psycopg2-binary
python-dotenv
requests
matplotlib
pyarrow
fastparquet